clean unused stuff
