+++ novawallet/Modules/AssetsSearch/AssetOperation/AssetOperationViewController.swift
@@ -1,8 +1,4 @@
 class AssetOperationViewController: AssetsSearchViewController {
-    var assetOperationPresenter: AssetOperationPresenterProtocol? {
-        presenter as? AssetOperationPresenterProtocol
-    }
-
     override func setupCollectionManager() {
         collectionViewManager = AssetOperationCollectionManager(
             view: rootView,
@@ -11,14 +7,4 @@ class AssetOperationViewController: AssetsSearchViewController {
             selectedLocale: selectedLocale
         )
     }
-}
-
-extension AssetOperationViewController: AssetOperationCollectionManagerDelegate {
-    func selectGroup(with symbol: AssetModel.Symbol) {
-        assetOperationPresenter?.selectGroup(with: symbol)
-    }
-}
-
-protocol AssetOperationCollectionManagerDelegate: AnyObject {
-    func selectGroup(with symbol: AssetModel.Symbol)
 }
